+++ src/main/java/com/travel/domain/diary/dto/response/AiDiaryResponse.java
@@ -1,0 +1,5 @@
+package com.travel.domain.diary.dto.response;
+
+public record AiDiaryResponse(
+        String diary
+) {}
