+++ src/main/java/com/travel/domain/diary/dto/response/DiaryResponse.java
@@ -1,0 +1,11 @@
+package com.travel.domain.diary.dto.response;
+
+import java.time.LocalDateTime;
+
+public record DiaryResponse(
+        LocalDateTime createdAt,
+        int resultCode,
+        String message,
+        String content,
+        Long diaryId
+) {}
